Revert "refactor markdown"
This reverts commit 6967a32d957aaa3a0c97a5417090a6707cd49e13.

Conflicts:
	app/models/article.rb
