--- app/models/markdown_preview.rb
@@ -1,11 +1,0 @@
-class MarkdownPreview
-  extend ActiveModel
-  include Markdownable
-
-  attr_accessor :body
-  markdownable :body
-
-  def initialize(body)
-    @body = body
-  end
-end
