+++ app/views/articles/show.html.erb
@@ -41,7 +41,7 @@
   </div>
 
   <div class="panel-body markdown">
-    <%= @article.body_html.html_safe %>
+    <%= markdown(@article.body).html_safe %>
   </div>
 </div>
 
@@ -66,7 +66,7 @@
           </span>
         </div>
         <div class="list-group-body markdown">
-          <%= comment.body_html.html_safe %>
+          <%= markdown(comment.body).html_safe %>
         </div>
         <% if comment.user == current_user %> 
           <%= render 'comments/form', hide: true, comment: comment %>
